some changes

+++ src/main/java/org/example/pricing/HourlyPricing.java
@@ -4,7 +4,7 @@ import org.example.vehicle.Vehicle;
 public class HourlyPricing implements PricingStrategy {
 
     @Override
-    public float calculatePrice(Vehicle vehicle, float hours) {
+    public float calculatePrice(Vehicle vehicle, int hours) {
         if (hours <= 0) return 0;
 
         return vehicle.getBasePricePerHour() * hours;
