+++ src/main/java/org/example/pricing/PricingStrategy.java
@@ -2,5 +2,5 @@ package org.example.pricing;
 import org.example.vehicle.Vehicle;
 
 public interface PricingStrategy {
-    float calculatePrice(Vehicle vehicle, float duration);
+    float calculatePrice(Vehicle vehicle, int duration);
 }
